Changes made RDCC-3854: Changes made to incorporate review comments
